rename pkg file->fs
