+++ pkg/fs/unique.go
@@ -1,4 +1,4 @@
-package file
+package fs
 
 import (
 	"fmt"
